[#4495] Add support for growing single-touch cache to a new overflow mark.
Summary:
Currently single-touch cache is given a fixed allocation of the total
cache(20%).
This change allows the cache to grow up to 80% of the whole allocation
when multi-touch entries are not present. We make sure that the
multi-touch cache always gets its whole 80% if it does have entries to
occupy that space by making we shrinking the single-touch cache.

This new overflow mark is configured by a new GFLAG:
cache_single_touch_overflow

Test Plan:
Added a new test case in CacheTest which makes sure that this behaviour
is ensured.

Reviewers: sergei, mikhail, bogdan, kannan, timur

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D8496